+++ CHANGELOG.md
@@ -408,3 +408,6 @@
 - Fixed a bug where the compiler would suggest to use a discarded value defined
   in a different function.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
+- Fixed a bug where the compiler allowed to write a guard with an empty clause.
+  ([Tristan-Mihai Radulescu][https://github.com/Courtcircuits])